refactor(github): decouple trellogithub and "github.com/google/go-github/v42/github" (#311)
closes #307
