+++ internal/pkg/plugin/trellogithub/github.go
@@ -2,15 +2,10 @@ package trellogithub
 
 import (
 	"fmt"
-	"net/http"
-	"strings"
 
-	githubx "github.com/google/go-github/v42/github"
 	"github.com/spf13/viper"
 
 	"github.com/merico-dev/stream/pkg/util/github"
-	"github.com/merico-dev/stream/pkg/util/log"
-	"github.com/merico-dev/stream/pkg/util/mapz"
 )
 
 // VerifyWorkflows get the workflows with names "wf1.yml", "wf2.yml", then:
@@ -24,7 +19,7 @@ func (tg *TrelloGithub) VerifyWorkflows(workflows []*github.Workflow) (map[strin
 	}
 
 	fmt.Printf("Workflow files: %v", wsFiles)
-	filesInRemoteDir, rMap, err := tg.FetchRemoteContent(wsFiles)
+	filesInRemoteDir, rMap, err := tg.client.FetchRemoteContent(wsFiles)
 	if err != nil {
 		return nil, err
 	}
@@ -33,40 +28,6 @@ func (tg *TrelloGithub) VerifyWorkflows(workflows []*github.Workflow) (map[strin
 	}
 
 	return tg.CompareFiles(wsFiles, filesInRemoteDir), nil
-}
-
-func (tg *TrelloGithub) FetchRemoteContent(wsFiles []string) ([]string, map[string]error, error) {
-	var filesInRemoteDir = make([]string, 0)
-	_, dirContent, resp, err := tg.client.Repositories.GetContents(
-		tg.ctx,
-		tg.options.Owner,
-		tg.options.Repo,
-		".github/workflows",
-		&githubx.RepositoryContentGetOptions{},
-	)
-
-	// error reason is not 404
-	if err != nil && !strings.Contains(err.Error(), "404") {
-		log.Errorf("GetContents failed with error: %s.", err)
-		return nil, nil, err
-	}
-	// StatusCode == 404
-	if resp.StatusCode == http.StatusNotFound {
-		log.Error("GetContents returned with status code 404.")
-		retMap := mapz.FillMapWithStrAndError(wsFiles, fmt.Errorf("not found"))
-		return nil, retMap, nil
-	}
-	// StatusCode != 200
-	if resp.StatusCode != http.StatusOK {
-		return nil, nil, fmt.Errorf("got some error: %s", resp.Status)
-	}
-	// StatusCode == 200
-	log.Info("GetContents return with status code 200.")
-	for _, f := range dirContent {
-		log.Infof("Found remote file: %s.", f.GetName())
-		filesInRemoteDir = append(filesInRemoteDir, f.GetName())
-	}
-	return filesInRemoteDir, nil, nil
 }
 
 // AddTrelloIdSecret add trello ids to secret
@@ -116,26 +77,4 @@ func (tg *TrelloGithub) AddTrelloIdSecret(trelloId *TrelloItemId) error {
 	}
 
 	return nil
-}
-
-func (tg *TrelloGithub) GetWorkflowPath() (string, error) {
-	_, _, resp, err := tg.client.Repositories.GetContents(
-		tg.ctx,
-		tg.options.Owner,
-		tg.options.Repo,
-		".github/workflows",
-		&githubx.RepositoryContentGetOptions{},
-	)
-
-	// error reason is not 404
-	if err != nil && !strings.Contains(err.Error(), "404") {
-		return "", err
-	}
-
-	// error reason is 404
-	if resp.StatusCode == http.StatusNotFound {
-		return "", err
-	}
-
-	return resp.Request.URL.Path, nil
 }
